+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/features/ShulkerMob.java
@@ -1,0 +1,46 @@
+package com.telepathicgrunt.repurposedstructures.world.features;
+
+import com.mojang.serialization.Codec;
+import net.minecraft.enchantment.Enchantments;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.ai.attributes.AttributeModifier;
+import net.minecraft.entity.ai.attributes.Attributes;
+import net.minecraft.entity.monster.ShulkerEntity;
+import net.minecraft.entity.monster.WitherSkeletonEntity;
+import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.util.Hand;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.ISeedReader;
+import net.minecraft.world.gen.ChunkGenerator;
+import net.minecraft.world.gen.feature.Feature;
+import net.minecraft.world.gen.feature.NoFeatureConfig;
+
+import java.util.Random;
+
+
+public class ShulkerMob extends Feature<NoFeatureConfig> {
+
+    public ShulkerMob(Codec<NoFeatureConfig> configFactory) {
+        super(configFactory);
+    }
+
+
+    @Override
+    public boolean generate(ISeedReader world, ChunkGenerator chunkGenerator, Random random, BlockPos position, NoFeatureConfig config) {
+        // move down to spawn at the jigsaw block calling this
+        position = position.down();
+
+        ShulkerEntity shulkerEntity = EntityType.SHULKER.create(world.getWorld());
+        shulkerEntity.enablePersistence();
+        shulkerEntity.setPosition(
+                (double)position.getX() + 0.5D,
+                position.getY(),
+                (double)position.getZ() + 0.5D);
+
+        shulkerEntity.setAttachmentPos(position);
+        world.spawnEntityAndPassengers(shulkerEntity);
+        return true;
+    }
+}